N337 Transferring skin code to skinning module
Just created a new Deform folder and added a new dpSkinning module to work with the skinning methods from there.
